update to work with layering
- standard is to copy the source segment buffer
- added option to copy "full stack" reading the last rendered frame: this should not be used with an overlapping "copy segment" as it copies itself (unless the user want that effect to happen)

the idea for full stack copy is:
- create a virtual steip/panel besides the physical LEDs (I used DDP with IP 0.0.0.0 as a test)
- Create a layered segment stack on the virtual strip/panel
- On the physical LED segment: copy FX of the virtual segment with "full stack" checked

The full-stack copy is not the cleanest most user-friendly solution. Its a bit convoluted to use and also lags one frame behind as it copies from the global frame buffer which is updated after all segments were rendered. Although not a perfect solution for a full-stack segment copy, I still think the option should be available: on larger setups with say 3-overlapping layers, a copy would require another 3 layers for each segment that the original is being copied to, eating up way more ram. AND: the option of just copying each segment is still available.