Update FX.cpp
I updated the checkboxes on Percent effect to support three new modes to automatically illuminate a % of the LEDs in a segment:
- % of minute
- % of hour
- % of day
This allows you to create dynamic effects based on these timescales. e.g. custom abstract clock displays. 
Personally I did this because I am creating a visual daily 'clock' for my kids with physical markers on it for "dad is done with work" and "Bed time" etc. I added minute/hour at the suggestion of others, but now this makes it suitable for true clock functionality which is nice -- and with much more flexibility than the other 'clock' mode option. 

This also deprecates the previous 'use single color' checkbox. I deemed this to be unnecessary since it is easy to just set the palette to 'color 1' which achieves the exact same effect.