FX.cpp: Update colorful for flexibility in # of colors from palette
Instead of hard-coding an override to use 5 colors for C9_2, add a function to automatically detect the number of colors in the palette (up to 16), so it can be used with custom palettes of more than 4 colors. Add a slider to manually specify number of colors, and change description of saturation slider to indicate its actual function (switching between "easter" colors, default colors, and the three color slots).