Set up Effect as a struct
Define an "Effect" struct and move the global list out of WS2812FX.
The important change is that the id code is now a property of the Effect
instead of an index in to a vector.

We require retrieving an ID code for an effect to call through a
function.  This allows us to change the back end implementation of the
numeric ID map in the future.

Segment::mode is preserved, but marked as deprecated.

Usermods not updated yet.