Remove OTA metadata structure version block
If we cap the acceptable version, it becomes impossible to increase it:
older firmwares will reject it.  Instead we must guarantee
backwards compatibility so long as the magic number is the same.
If a breaking change is required in the future, a new magic number
should be used to identify it.